# devops-final-be/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the DevOps demo project.",
        "title": "DevOps Demo API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/api/admin/audit-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all audit logs. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Get all audit logs",
                "responses": {
                    "200": {
                        "description": "Audit logs retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "type": "array"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing user ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/audit-logs/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search audit logs by table name, record id, user id, action, old value, or new value. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Search audit logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search text",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit logs retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "type": "array"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing user ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/audit-logs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single audit log by ID. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Get audit log by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit Log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit log retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {
                                    "type": "object"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing user ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Audit log not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "result": {},
                                "status": {
                                    "type": "string"
                                },
                                "status_code": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}